Complete Revamp
Completely re-did most of the code. Now the collision of the blocks is done with precise calculations separately from the visual generation. Blocks are visually painted where they are estimated to be. The program works, but past a weight of 10000 graphical glitches arise and the time taken for the last few collisions is very long. More tuning needed.